% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_stock.R
\name{download_stock}
\alias{download_stock}
\title{download_stock}
\usage{
download_stock(symbol, source, period = 500)
}
\arguments{
\item{symbol}{stock symbol such as "AAPL"}

\item{source}{where is the data from, "yahoo" for Yahoo Finance or "tws" for
for Interactive Brokers TWS. Need a account and data subscription for TWS.}

\item{period}{number of days to look back}
}
\value{
data frame
}
\description{
Download price data of a stock
}
