% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_momentum.R
\name{plot_momentum}
\alias{plot_momentum}
\title{Follow short term trend after a price jump}
\usage{
plot_momentum(
  stocks,
  source = "tws",
  ema1 = 5,
  ema2 = 20,
  ema3 = 200,
  macd1 = 12,
  macd2 = 26,
  macd3 = 9,
  period = 500,
  zoom = 0.3
)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}

\item{source}{where is the data from, "yahoo" for Yahoo Finance or "tws" for
for Interactive Brokers TWS. Need a account and data subscription for TWS.}

\item{ema1}{integer, short exponential moving average (EMA)}

\item{ema2}{middle EMA}

\item{ema3}{long EMA}

\item{macd1}{integer, MACD fast MA}

\item{macd2}{MACD slow MA}

\item{macd3}{MACD signal}

\item{period}{number of days to look back}

\item{zoom}{numeric scale of the MACD to display}
}
\description{
Follow short term trend after a price jump
}
