% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_momentum.R
\name{plot_momentum_}
\alias{plot_momentum_}
\title{Plot momentum of a stock}
\usage{
plot_momentum_(
  symbol,
  source = "tws",
  ema1 = 5,
  ema2 = 20,
  ema3 = 200,
  macd1 = 12,
  macd2 = 26,
  macd3 = 9,
  period = 500,
  zoom = 0.3
)
}
\description{
Plot momentum of a stock
}
