% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_momentum.R
\name{plot_momentum}
\alias{plot_momentum}
\title{Follow short term trend after a price jump}
\usage{
plot_momentum(
  stocks,
  source = "tws",
  ema1 = 5,
  ema2 = 20,
  ema3 = 200,
  macd1 = 12,
  macd2 = 26,
  macd3 = 9,
  period = 500,
  zoom = 0.3
)
}
\arguments{
\item{stocks}{vector of stock symbols such as c("C", "AAPL")}

\item{period}{number of days to look back}

\item{description}{description of the chart, stock symbol by default}
}
\description{
Follow short term trend after a price jump
}
